+++ codemp/qcommon/stringed_ingame.cpp
@@ -667,7 +667,7 @@ LPCSTR CStringEdPackage::ParseLine( LPCSTR psLine )
 					psWordEnd++;
 				}
 				char sThisLanguage[1024]={0};
-				int iCharsToCopy = psWordEnd - psLine;
+				size_t iCharsToCopy = psWordEnd - psLine;
 				if (iCharsToCopy > sizeof(sThisLanguage)-1)
 				{
 					iCharsToCopy = sizeof(sThisLanguage)-1;
@@ -765,7 +765,7 @@ LPCSTR Leetify( LPCSTR psString )
 						};
 
 		char *p;
-		for (int i=0; i<sizeof(cReplace); i+=2)
+		for (size_t i=0; i<sizeof(cReplace); i+=2)
 		{
 			while ((p=(char*)strchr(str.c_str(),cReplace[i]))!=NULL)
 				*p = cReplace[i+1];
@@ -1047,7 +1047,7 @@ int SE_GetNumFlags( void )
 
 LPCSTR SE_GetFlagName( int iFlagIndex )
 {
-	if ( iFlagIndex < TheStringPackage.m_vstrFlagNames.size())
+	if ( iFlagIndex < (int)TheStringPackage.m_vstrFlagNames.size())
 	{
 		return TheStringPackage.m_vstrFlagNames[ iFlagIndex ].c_str();
 	}
@@ -1123,7 +1123,7 @@ int SE_GetNumLanguages(void)
 //
 LPCSTR SE_GetLanguageName( int iLangIndex )
 {
-	if ( iLangIndex < gvLanguagesAvailable.size() )
+	if ( iLangIndex < (int)gvLanguagesAvailable.size() )
 	{
 		return gvLanguagesAvailable[ iLangIndex ].c_str();
 	}
@@ -1136,7 +1136,7 @@ LPCSTR SE_GetLanguageName( int iLangIndex )
 //
 LPCSTR SE_GetLanguageDir( int iLangIndex )
 {
-	if ( iLangIndex < gvLanguagesAvailable.size() )
+	if ( iLangIndex < (int)gvLanguagesAvailable.size() )
 	{
 		return va("%s/%s", sSE_STRINGS_DIR, gvLanguagesAvailable[ iLangIndex ].c_str() );
 	}
